* moved db package into utils.
