+++ cmd/committer/db_init_cmd.go
@@ -15,7 +15,7 @@ import (
 
 	"github.com/hyperledger/fabric-x-committer/cmd/cliutil"
 	"github.com/hyperledger/fabric-x-committer/cmd/config"
-	"github.com/hyperledger/fabric-x-committer/service/db"
+	"github.com/hyperledger/fabric-x-committer/utils/db"
 )
 
 const initDBCommand = "init-db"
