+++ utils/db/create_namespace_tmpl.sql
