+++ utils/db/init_database_tmpl.sql
